clio-agent-obit/requirements.txt
python-dotenv>=1.0
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
python-gedcom>=1.0

# Valfria:
# feedparser>=6.0           (för deprecated RSS-adapter, source_familjesidan_rss.py)
# python-Levenshtein>=0.12  (snabbare + bättre namnmatchning, faller tillbaka på difflib)

# clio-core installeras separat:
# pip install -e ../clio-core
